extract pagination state computation
